build: replace `ngx-swiper-wrapper` with `swiper/angular`
As mentioned in the `ngx-swiper-wrapper`'s [README][1], the
`ngx-swiper-wrapper` project is not going to have another release,
because the latest version of the official `swiper` package has built-in
[support for Angular][2].

This commit removes the `ngx-swiper-wrapper` dependency (which is not
compatible with the `swiper` v7) and switches to using `swiper`
directly.

[1]: https://github.com/zefoy/ngx-swiper-wrapper#eol-notice
[2]: https://swiperjs.com/angular
