+++ package.json
@@ -250,7 +250,6 @@
     "ngx-simple-modal": "1.4.17",
     "ngx-smart-modal": "7.4.1",
     "ngx-spinner": "12.0.0",
-    "ngx-swiper-wrapper": "10.0.0",
     "ngx-toastr": "14.1.0",
     "ngx-ui-loader": "11.0.0",
     "ngx-wig": "12.2.0",
@@ -265,7 +264,7 @@
     "reflect-metadata": "0.1.13",
     "rxjs": "7.3.0",
     "rxjs-compat": "6.6.7",
-    "swiper": "6.8.4",
+    "swiper": "7.0.2",
     "tns-core-modules": "6.5.27",
     "ts-helpers": "1.1.2",
     "tslib": "2.3.1",
