+++ projects/swiper-ngcc/e2e/src/app.e2e-spec.ts
@@ -10,7 +10,7 @@ describe('workspace-project App', () => {
 
   it('should display welcome message', () => {
     page.navigateTo();
-    expect(page.getTitleText()).toEqual('Welcome to ngx-swiper-wrapper-ngcc!');
+    expect(page.getTitleText()).toEqual('Welcome to swiper-ngcc!');
   });
 
   afterEach(async () => {
