+++ projects/swiper-ngcc/src/app/app.component.ts
@@ -6,5 +6,5 @@ import { Component } from '@angular/core';
   styleUrls: ['./app.component.css']
 })
 export class AppComponent {
-  title = 'ngx-swiper-wrapper-ngcc';
+  title = 'swiper-ngcc';
 }
