+++ projects/swiper-ngcc/src/app/app.module.ts
@@ -1,0 +1,17 @@
+import { BrowserModule } from '@angular/platform-browser';
+import { NgModule } from '@angular/core';
+import { SwiperModule } from 'swiper/angular';
+
+import { AppComponent } from './app.component';
+
+@NgModule({
+  declarations: [
+    AppComponent
+  ],
+  imports: [
+    BrowserModule,
+    SwiperModule,
+  ],
+  bootstrap: [AppComponent]
+})
+export class AppModule { }
