+++ projects/swiper-ngcc/src/favicon.ico
