+++ projects/swiper-ngcc/src/polyfills.ts
