+++ projects/swiper-ngcc/tslint.json
