+++ tsconfig.json
@@ -1442,13 +1442,13 @@
       "path": "./projects/ngx-spinner-ngcc/e2e/tsconfig.json"
     },
     {
-      "path": "./projects/ngx-swiper-wrapper-ngcc/tsconfig.app.json"
+      "path": "./projects/swiper-ngcc/tsconfig.app.json"
     },
     {
-      "path": "./projects/ngx-swiper-wrapper-ngcc/tsconfig.spec.json"
+      "path": "./projects/swiper-ngcc/tsconfig.spec.json"
     },
     {
-      "path": "./projects/ngx-swiper-wrapper-ngcc/e2e/tsconfig.json"
+      "path": "./projects/swiper-ngcc/e2e/tsconfig.json"
     },
     {
       "path": "./projects/ngx-toastr-ngcc/tsconfig.app.json"
